+++ src/components/layout/header/BackHeader.tsx
@@ -1,5 +1,5 @@
-import BaseHeader from './BaseHeader';
-import { BackButton } from './Header';
+import BaseHeader from './base/BaseHeader';
+import { BackButton } from './base/Header';
 
 const BackHeader = ({ onBackClick }: { onBackClick: () => void }) => (
   <BaseHeader leftSection={<BackButton onClick={onBackClick} />} />
