+++ src/components/layout/header/LogoHeader.tsx
@@ -1,5 +1,5 @@
-import BaseHeader from './BaseHeader';
-import { Logo } from './Header';
+import BaseHeader from './base/BaseHeader';
+import { Logo } from './base/Header';
 
 const LogoHeader = () => <BaseHeader leftSection={<Logo />} />;
