+++ src/components/layout/header/LogoSettimgHeader.tsx
@@ -1,5 +1,5 @@
-import BaseHeader from './BaseHeader';
-import { Logo, SettingsButton } from './Header';
+import BaseHeader from './base/BaseHeader';
+import { Logo, SettingsButton } from './base/Header';
 
 const LogoSettingsHeader = ({ onSettingsClick }: { onSettingsClick: () => void }) => (
   <BaseHeader leftSection={<Logo />} rightSection={<SettingsButton onClick={onSettingsClick} />} />
